+++ .env.template
@@ -1,0 +1,3 @@
+OPENAI_API_KEY=
+LANGCHAIN_API_KEY=
+EMBEDDINGS_MODEL='text-embedding-3-small'
