--- prototipo/.env.template
@@ -1,2 +1,0 @@
-OPENAI_API_KEY=
-GROQ_API_KEY=
