+++ requirements.txt
@@ -5,5 +5,6 @@ langchain_core
 langchain_experimental
 langchainhub
 python-dotenv
+faiss-cpu
 
 streamlit==1.34.0
